hotel_management/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
python-dotenv==0.19.0
pycryptodome==3.10.1
Flask-WeasyPrint==0.6
requests==2.26.0